requirements.txt
aiohttp==3.7.2
alabaster==0.7.12
anyjson==0.3.3
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.3.0
aspy.refactor-imports==2.1.1
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
autoflake==1.4
Babel==2.8.0
bandit==1.6.2
black==20.8b1
cached-property==1.5.2
cent==3.0.1
certifi==2020.6.20
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
cycler==0.10.0
da-vinci==0.3.0
distlib==0.3.1
Django==3.1.3
django-admin-generator==2.2.0
django-admin-tools==0.9.0
django-appconf==1.0.4
django-cors-headers==3.5.0
django-print-sql==2018.3.6
django-rosetta==0.9.5
django-thumbnails==0.3.2
django-unused-media==0.2.2
django-user-agents==0.4.0
django-utils2==2.8.0
django-watch==0.4
djangorestframework==3.12.2
docutils==0.16
eradicate==2.0.0
feedparser==6.0.2
filelock==3.0.12
flake8==3.8.4
flake8-annotations==2.4.1
flake8-annotations-coverage==0.0.4
flake8-bandit==2.1.2
flake8-broken-line==0.3.0
flake8-bugbear==20.1.4
flake8-builtins==1.5.3
flake8-commas==2.0.0
flake8-comprehensions==3.3.0
flake8-debugger==3.2.1
flake8-eradicate==1.0.0
flake8-executable==2.0.4
flake8-fixme==1.1.1
flake8-future-import==0.4.6
flake8-mutable==1.2.0
flake8-plugin-utils==1.3.1
flake8-polyfill==1.0.2
flake8-pyi==20.10.0
flake8-pytest==1.3
flake8-pytest-style==1.3.0
flake8-string-format==0.3.0
flake8-todo==0.7
flake8-unused-arguments==0.0.6
geoip2==4.1.0
gitdb==4.0.5
GitPython==3.1.11
icmplib==2.0.1
identify==1.5.9
idna==2.10
imagesize==1.2.0
iniconfig==1.1.1
isort==5.6.4
Jinja2==2.11.2
jsonschema==3.2.0
jsonschema-to-openapi==0.2.1
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
mailjet-rest==1.3.4
MarkupSafe==1.1.1
maxminddb==2.0.3
mccabe==0.6.1
multidict==5.0.0
mypy-extensions==0.4.3
nodeenv==1.5.0
numpy==1.19.4
openapi-schema-pydantic==1.1.0
packaging==20.4
pathspec==0.8.0
pbr==5.5.1
Pillow==8.0.1
ping3==2.6.6
pluggy==0.13.1
polib==1.1.0
pre-commit==2.8.2
prettytable==2.0.0
psycopg2-binary==2.8.6
py==1.9.0
pycodestyle==2.6.0
pydantic==1.7.2
pyflakes==2.2.0
Pygments==2.7.2
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.1.2
pytest-django==4.1.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-memcached==1.59
python-slugify==4.0.1
python-telegram-bot==13.3
python-telegram-bot-calendar==1.0.5
python-utils==2.4.0
python-watch==0.1
pytils==0.3
pytz==2020.4
pyupgrade==2.7.3
PyYAML==5.3.1
regex==2020.10.28
reorder-python-imports==2.3.6
requests==2.24.0
selenium==3.141.0
sentry-sdk==0.19.2
sgmllib3k==1.0.0
shortuuid==1.0.1
simple-print==0.5
six==1.15.0
smmap==3.0.4
snowballstemmer==2.0.0
sorl-thumbnail==12.6.3
Sphinx==3.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-redoc==1.6.0
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
stevedore==3.2.2
tcping==0.1.1rc1
termcolor==1.1.0
text-unidecode==1.3
tokenize-rt==4.0.0
toml==0.10.2
tornado==6.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
ua-parser==0.10.0
ujson==4.0.1
urllib3==1.25.11
user-agents==2.2.0
virtualenv==20.1.0
wcwidth==0.2.5
webbot==0.34
wrapt==1.12.1
yarl==1.6.2
yesqa==1.2.2
gunicorn 
shortuuid